kernel: fix conflict function bugs with ta > 0.0
If the external events have the priority over internal event, the
internal event was invalided and deleted by the newly internal event
build after external transition and pushed into the scheduler. If this
internal event was scheduled to a date greater than 0, the internal
event at current date is not called.