vle: fix crash when model deletion and timed observation
If a model is deleted into a bag, and an observation of this model is
scheduled in the next bags (minimum, next second bag), then a segfault
occurred. The model is deleted into the next bag of the deletion, not at
the end of the complete event bags (all bags at the same date).