vle: fix crash when cloning classes
The cloning process for classes is buggy. When the hierarchy of DEVS
models is cloned, observables are attached too quickly to the stream,
ie. before the attachment to the parent. Stream plug-ins throw error
because they receive multiples 'addNewObservable' event with the same
model name. Fix-it by attaching the cloned class to the parent before
creating dynamics. (closes: #188).