oov: Add new plugin - netview.
The netview plugin enables to visualize graphs/networks. Each node is
created with the name of the corresponding devs atomic model it is
associated with or automatically named with a given prefix. Position of
nodes can be defined by the user or automatically generated. The plugin
enables to display dynamic graph/networks and vertices/nodes properties
as changing colors.

The "gens" model is modified to provide an example of using.