oov: remove useless observation in SimpleFile plug-ins
By default, the SimpleFile plug-ins (csv, text, rdata) add a line when a
simulator is attached to the view with a Null (or NA) value. If no model
sends data (observations), the line contains only NA. For instance, it
produces:

"time" "top:X.x" "top:Y.y"
0 NA NA
1000 2550 1

We fix it by adding a mask before writing the line. If the mask is not
empty, the line is not writing (closes: #195).