gvle: Add feature in AtomicModelBox to rename ports
Addition of the function onRename in the InputPortTreeView class, based on the
functions onAdd and onRemove, It calls the function renameInputPort (in the
class Model). It is a mix between the functions addInputPort and delInputPort.
It saves the connection in relation with the port, delete the port and create a
new port with the saved connections (closes: #2722148).