gvle: reactivate the classes of model
Classes of models are models which are not present in the experiment.
gvle/ModelClassBox : The buttons are removed and a menu is available
on right clic (ClassTreeView), to add, delete and rename the
class.
The edition is available with the double-clic.
If the model is coupled, a view is opened.
If the model is atomic, the atomic model box is opened.
gvle/Modeling :
* Reactivation of the function getClassModel to recover the model
  associated to the class.
* Creation of a function setModelClass to associate a model and a class.
* Creation of the functions addViewClass which create a new view in
  related with the models' class.
* Creation of a function addAtomicModelClass to simplify the addition
  of an atomic model into a class.
* Creation of a function getAtomicModelClass to return the AtomicModelList
  associated to a class.
* Modification of the function cut, copy, paste and delModel to check
  if it is a Class's name.
gvle/NewModelBox : the function on_apply can create a new class.
gvle/View :
* Creation of an attribute mCurrentClass, empty if it is
  not a class, or containing the name of the class.
* Verification before called the function addView if the model belongs
to a class with the attribute mCurrentClass.
- Addition of a parameter to the functions cut, copy, paste and delModel
vpz/Class : Addition of a function setAtomicModel to copy an
  AtomicModelList.
(closes: #2723012)

Signed-off-by: Patrick Chabrier <patrick.chabrier@toulouse.inra.fr>