vle: fix a bad reading of seed attributes
The attributes 'seed', in experiment and replicas xml tags, needs to be
encoded in an unsigned int 32 bits (see boost mersene twister API). The
sax parser read it like a long int and can throw an error on 32
architecture (sizeof(int) equal sizeof(long). We fix it by adding an
xmlCharToUnsignedInt function and read seed with this function.
(Closes: #2819683).