utils: add template file system
To allow the development of template plug-in (ie. generation of source
code), we need a system based on C++ template files filled by template
variable like in web template service. The system needs to provide
instruction "name", "if" and "for". (Closes: #2866878).

For instance, the differenceequation.template.file:

	class {{name}} : public DifferenceEquation
	{
	public:
		{{name}}(...) {}

		virtual ~{{name}} {}

		{{if declare}}
		{{for i in list}}
		var {{list^i}};
		{{end for}}
		{{end if}}
	}:

In a pseudo GVLE template plug-in:
	[...]
	vle::utils::Template tpl;
	tpl.open("differenceequation.template.file");
	tpl.stringSymbol("name", "meteo");
	tpl.listSymbol("list", "x");
	tpl.listSymbol("list", "y");
	tpl.listSymbol("list", "z");
	tpl.boolSymbol("declare", true);
	tpl.process(std::cout);
	[...]