vle: package command check process status
The 'package::build' method uses 'make all' and 'make install' commands,
but the 'make install' command runs 'make all' if all targets are not
available. We fix the problem by adding a control of the status of spawned
process on Unix and Win32 and check the result of 'make all' before
starting the 'make install' (Closes: #2856124).

* Add a Linux/Win32 codes to control status of spawned process.
* Update the command line interface to stop the process if a error appear
  in a package subprocesses.
* Fix a typo copy/paster error in GVLE package command which called
  'clean()' instead of 'pack()'.