gvle: fix crash when renaming file
The file list treeview in the GVLE GUI allows user to rename files and
directories with an editable column, but no test are performed on the new
filename and on the rename operation. This behaviour can destroy files. We
fix it by removing the editable column and replace it by a static content.