gvle: add method to set graphical attributes
When the graphical attributes of models are not defined in VPZ file, the
display in graphical view is wrong and it's not possible to modify
graphically the coordinates. We have add a new method to define
automatically x, y, width and height. This new method is accessible with
random placement procedure (menu View -> Random order) (Closes: #2862639).

Signed-off-by: Gauthier Quesnel <quesnel@users.sourceforge.net>