gvle: fix bad packages name if OpenPackageBox
The boost::filesystem::basename does not return the filename of a path,
but the filename without extension
(boost::filesystem::basename("/tmp/x.log") returns "x"). The call of
this function introduces a bug in OpenPackageBox which can not open
packages with dot in their name.

Add an utils::Path::filename function that returns the filename of a
path (with extension) and we use this function in OpenPackageBox and for
the GVLE files hierarchy. (Closes: #2883848).