ext/dec: fix output function of the Agent
The state Output may be canceled if the Agent model receives an external event
after an internal or an external transition. In this case, the time advance
function may returns a bad duration. We correct the problem by calling output
function if at least one latest activities lists is not empty.

* Add an unit test.