ext/de: fix bug on dynamic creation and history
When a coupled model is created dynamically and it have several
atomic models based on DifferenceEquation extension, if an atomic
model have a history, this history goes out the coupled model and
it is possible that an atomic model is running and receive this
history. This case proceed a bug (an exception).