gvle: fix the setting of double values
When a vpz contained values with many digits, they were not displayed
properly in the entries.
We added a method in Utils that manage the number of digit.
The method enable the use of '.' as separator in GVLE.
Values can be set like this 3,14e-18 or like that 3.14e18.

* modify the "Simulation/Project" and "Simulation/view" form to handle
  scientific values with many digits.
* add a method that manages the number of digit and detect the english
  number notations.
* add unit tests for convertion.
(closes: #2983669)

Signed-off-by: Patrick Chabrier <patrick.chabrier@toulouse.inra.fr>