gvle: add a dynamics list editor
A new dedicated class (DynamicsBox) to manage dynamics is added. A
specific window with similar behaviour as the Output Views editor is
provided
This new window can be used to :
* Create a new dynamic
* Edit a dynamic (package, library, model, location and language)
* Copy a dynamic
* Remove a dynamic
* Rename a dynamic
The previous way to manage dynamics is still fully working. And from
the new editor the access to the modelisation plugin is not yet
available.
(closes: #2996470)

Signed-off-by: Patrick Chabrier <patrick.chabrier@toulouse.inra.fr>