vle: fix conflict in cli with multiple vpz
When starting a simulation with a vpz (in CLI), if a vpz file exists in
the current directory and a file exists in exp directory of the package,
by default, the priority is set on the current file. We change the
behavior by throwing a exception in this case.

- Add a function utils::Path::getCurrentPath().
- Update the behavior of the CLI.
(cherry picked from commit a3311c373611c7bb915b2fcc778c44f891619013)