gvle: update the selection of dynamic and observable
A column of radio button is added inside the treeviews of dynamics and
observables to provide a standard single row selection system. From
now on, selecting a row does not mean choosing it.
A sort is enabled for all the columns of each treeviews of the
AtomicModelBox.
The setting of the sort mode on columns is simplified.
(closes: #3013774)

Signed-off-by: Patrick Chabrier <patrick.chabrier@toulouse.inra.fr>