gvle: add file copy and paste functionnality
In order to Copy/Paste files, two entries have been added to the
package treeview. If the file already exist in the target folder, it
is automatically renamed as "file_1.ext", and the file file name is
directly proposed to be edited. This can be concevient wants to change
the name.  The possibility to rename a file by clicking on the name is
added.  The file extension can not be changed in this treeview.
(closes: #3009527)

Signed-off-by: Patrick Chabrier <patrick.chabrier@toulouse.inra.fr>