gvle: fix the control of the plugins C++ identifier
In the Statechart plugin, it was possible to write wrong alphanumeric
string according to the C++ identfier rule, and also for the 3
parameters of one state.  Now if the parameter is not compatible with
the name of a C++ identfier, the validate button and the source button
are not sensitive.
(closes: #2996670)

Signed-off-by: Patrick Chabrier <patrick.chabrier@toulouse.inra.fr>