gvle: fix the action state name entry
When usind the Statechart plugin, the problem happens if the user
doesn't edit a method correpondind to the parameters InAction,
OutAction or Activity, but only specify the name of these actions. In
this case the name was not any more available for change at the next
session.

The ComboBoxEntryText list is fed with the list of available
action. And the ComboBoxEntryText entry is preset with a action name
not available inside the list. This is coherent while the
correspondind method is not available yet.
(closes: #3009524)

Fix the wrong reedition behaviour of the IN and Out actions. Now they
can be reedited. Therefore the "action" accessor is used insteed of
the "eventAction" one.

Signed-off-by: Patrick Chabrier <patrick.chabrier@toulouse.inra.fr>