oov: add a on-the-fly output stream
By default, the oov file plug-in can only use standard output at the end
of the simulation. This approach cancels the possibility of using tools
which draws on-the-fly for instance Gnuplot with scripts. We add a new
oov plug-in which flush observation directly on the standard output
stream. The last line is used to show the column's name.

Parameters are:
- locale: used C by default, but can use the user's locale (`user') or a
  specified locale (fr_FR).
- julian-day: to show a second column of time with a conversion of the
  time into a julian date.
- header: to show (by default, header is hide) in the latest line of the
  observation.