gvle: fix the Gtk::TextView log scroll
The default behavior of the Gtk::TextView log (shows the results of
configure, build, test, clean and package commands) is not effective. It
does not scroll across the entire buffer. This patch uses the
Gtk::TextBuffer::Mark to store the last line changed and uses this Mark
to scroll effectively the Gtk::TextView.

- add an effective scroll functionality.
- use insertLog function to reduce the source.