devs: fix bad id in ModelFactory
The devs::ModelFactory uses the name of the library to distinguish
loaded Glib::Module. However, the name of the library is not sufficient
now with the package mode introduced in VLE 0.8. We fix algorithm and
ModelCache to use the Glib::Module::get_name() function instead of the
vpz library string.

- Update the devs::ModelFactory and devs::ModelCache.
- The devs::ModelFactory::getPlugin is deprecated.

- Optimize the devs::Coordinator::addModel function 0(2log(n)) to O(log(n)).