cli: add a make all command in cli
Add a new command in the CLI to build all depends for the current
package. The function list for all vpz in the experiment directory the
list of depends (in dynamics or outputs plugins) and try to configure
and build each of them. For example:

goth@daviu> vle -P archidemio all
[...]