gvle: add enhancement of entries in modeling plugins
Some dialogs of the modeling plugins fsa and petrinet have been
enhanced. Information stored after generation are used another way.
Wen a name of a method is set by the user, even if the C++ definition
is not provided, the name is available the next time the user wants to
change it. The dialogs where this behaviour is available:
- EventInStateDialog of the fsa plugin
- TransitionDialog of the fsa plugin
Wen a name of a port is set by the user, even if not defined on the model,
the name is available the next time the user wants to change it. The
dialogs where this behaviour is available:
- PlaceDialog of the petrinet plugin
- TransitionDialog of the petrinet plugin