gvle: add two commands to the Project menu
The "Project Dependencies" entry of the Project menu allow to show the
list of all the dependecies in terms of package of the current package
by launching a dedicated information window. The window display only
the first level of dependependencies for each simulator.
- Add(duplicate) the depends function that gives the complete list of
  dependencies to GVLE.
- Add the displayDependencies function that format the complete list
  of dependencies and launch a MessageDialog to GVLE.

The "Build All Entries" entry of the Project menu allow to launch a
"configure, build, install" workflow on all the dependencies of the
current package and on himself also. This command only manage the
first level of dependependencies.
- Add the function makeAllProject to GVLE, in charge to get the
  dependencies, and to launch the workflow.
- Add the buildAllProject, configureAllProject, installAllProject
  functions to GVLE. Each of them launch a convenient timer, in order
  to be inserted in the workflow.
- Add the timer packageAllTimer to GVLE. This timer can either
  continue the worflow if another package is to treat or stop.
- Add the timer packageBuildAllTimer to GVLE . This Timer can continue
  the worflow by calling the install command on the current package.
- Add the timer packageConfigureAllTimer to GVLE . This Timer can
  continue the worflow by calling the build command on the current
  package.

cli: update the all command. The package of interest is built also.