gvle: fix the modeling plug-in launching
The code generation and the code parsing was note done accordingly to
the new plug-in management system. From the AtomicModelBox, a modeling
dialog on an already dynamic could not be launched anymore. Each
template of each plug-in has been updated. The template tag parser has
been enhanced to read the new plug-in name convention, where the name of
the package is stored inside brackets.

Signed-off-by: Gauthier Quesnel <quesnel@users.sourceforge.net>