devs: fix crash when throwing exception in model's
Since the update of the value::Map class, the experimental conditions
are not cloned but just referenced. Thus, we must to clear the
experimental condition data instead of deleting the data causing a
double free memory. We fix the problem by clearing the value::Map
instead of deleting it(closes: #3288773, #3184804).