utils: fix crash with executive and wrapper dynamics
The ModuleManager does not propagate correctly the ModuleType
(MODULE_DYNAMICS, MODULE_DYNAMICS_EXECUTIVE OR MODULE_DYNAMICS_WRAPPER)
which is use by the devs::ModelFactory. We fix the ModuleManager by
assigning the ModuleType in the public API ModuleManager::get function
with a new API:

    void *get(const std::string& package, const std::string& library,
              ModuleType type, ModuleType *newtype = 0) const;

This function now returns in @e newtype parameter the type of the module
read. The @e type is MODULE_DYNAMICS, MODULE_DYNAMICS_EXECUTIVE or
MODULE_DYNAMICS_WRAPPER the @e newtype check the type and if it does not
corresponds it check the two other simulators (MODULE_DYNAMICS,
MODULE_DYNAMICS_EXECUTIVE or MODULE_DYNAMICS_EXECUTIVE). If @e type is
not MODULE_DYNAMICS, MODULE_DYNAMICS_WRAPPER, MODULE_DYNAMICS_EXECUTIVE,
the @e newtype will be affected by the founded type.

     // $VLE_HOME/.vle/pkgs/glue/lib/libcounter.so
     ModuleManager manager;
     ModuleType type;
     manager.get("glue", "counter", MODULE_DYNAMICS, &type);
     assert(type == MODULE_DYNAMICS || type == MODULE_DYNAMICS_WRAPPER ||
            type == MODULE_DYNAMICS_WRAPPER);

     manager.get("glue", "counter", MODULE_DYNAMICS_WRAPPER, &type);
     assert(type == MODULE_DYNAMICS || type == MODULE_DYNAMICS_WRAPPER ||
            type == MODULE_DYNAMICS_WRAPPER);

     manager.get("glue", "counter", MODULE_OOV, &type);
     assert(type == MODULE_OOV);