manager: remove the seed and replicas
To prepare the next kernel of VLE we remove global object such as the
pseudo random generator (PRNG) from the coordinator and referenced in
all classes which inherit of vle::devs::Dynamics. We remove also the
manager PRNG which can not be used by models and consequently, we change
the OutputSimulationMatrix into OutputSimulationList to store result of
a Manager execution.