gvle/modeling: fix the Gtk::Entry reset bug
As soon as a Gtk::Entry was reset to an empty string, on the windows
portage, an error occurs. There is 3 way to fix this, not to reset
entries, reset entries with a dummy string or reset to an empty string
directly the buffer. We decide to use first the first policy and
secondly the third policy.
(Closes: #3448080)