utils: change $VLE_HOME/pkgs into $VLE_HOME/pkgs-x.y
To allow the use of VLE 1.0 and 1.1 on the same $VLE_HOME directory, we
need to clearly split the packages from the two versions. We add in the
VLE's version.hpp file, a macro VLE_ABI_VERSION equal to MAJOR.MINOR. We
use this macro to define the name of the package directory. For example,
in VLE 1.1, the package directory is defined as vle/pkgs-1.1 and in VLE
1.2, the package directory is defined as vle/pkgs-1.2. The current
stable version of VLE is not change and use packages in vle/pkgs
directory.