utils: move package unit test
In test1.cpp we check packages system by building the `tmp' and `x'
packages. But, to perform the tests, we install these package into the
users' home directory ($VLE_HOME) and we not remove it.

To fix the problem, we move the unit tests to the test_package.cpp. In
this file, the $VLE_HOME environment variable is override to the
temporary directory of the system.