cli: checking existence of package before select
In cli, the VLE_HOME/pkgs-1.1/package/output directory is build by
default even if the package does not exist and without create package
command.

For example, if the toto package does not exist, the configure command
fail but the directory toto is build.

$ vle-1.1 -P toto configure

We fix the cli via a best test. (closes #43).