win32: fixing Gtk problem with dll and application
With mingw32, gvle.exe (only depends to libgvle.dll and libvle.dll), the
Gtk::Window is not displayed if the Gtk::Main loop is in gvle.exe and the GVLE
class is in the libgvle.dll. This seems to be a bug in Gtk since with have no
problem to run gvle.exe statically linked with libgvle. To fix the problem, we
introduce the static start function in libgvle.dll which wraps the Gtk::Main
loop into libgvle.dll. Now the gvle.exe application have only a call to this
function.