devs: remove EventList class
We remove the vle::devs::EventList class and the EventList.hpp file.
This class is just an std::vector with addEvent and deleteAllAndClear
functions. We replace the class vle::devs::ExternalEventList with a
typedef equal to:

	std::vector < vle::devs::ExternalEvent* >

We upgrade source code in vle::devs namespace to take into account this
change.

This patch breaks API of vle::devs::Dynamics externalTransition,
confluentTransition and output functions (closes #54).