devs: add a runtime access to the simulation results
A runtime access to the results has been provided to the RootCoordinator
class via the outputs function.

Be careful, to avoid memory leak, user need to delete newly allocated
value::Map after the deletion of the RootCoordinator. Ok, it's a bad
design but we need to conserve API.

Example:

int main()
{
    vle::value::Map *result;

    {
        vle::utils::ModuleManager mm;
        vle::devs::RootCoordinator rc(mm);
        ...
        result = rc.outputs();
    }

    delete result; // avoid memory leaks.
}