cvle: add a new cluster simulation tools
CVLE is a program to execute a set of simulation on classical cluster
system: MPI (Message Parsing Interface).

CVLE treats huge simulation plan. CVLE does not use the manager to build
the experimental frames. It reads a Comma-Separated Value  (CSV)  where
columns  are conditions,  lines  are  instances  of the experimental
frame. The first line defines the conditions. If a column does not exist
in the VPZ file, it is simply added to he output file (e.g. to store
simulation id). Input and output of the simulation use only simply type:
boolean, integer, double and string.