vle: remove stdlocale call in Init()
The vle::Init::Init() constructor should not call the setlocale
function.

If a user needs to set locale for the VFL (as in Gvle, Cvle, Mvle) we
provide Init::Init(const char *localename) function. If the user needs
to call setlocale in its own code (like in rvle package), we provide
Init::Init() constructor (Closes #219).