devs: updates data in Observation and External events
We replace the boost::shared_ptr<value::Map> with a
generic std::shared_ptr<value::Value> and we update the API to take into
account this change.

We remove the useless value::Map in ObservationEvent class.

Therefore, we remove the vle::devs::Attribute class and Attributes.hpp,
ObservationEvent.cpp files.