devs: VFL plug-ins can be load in program instead of package
From now, ModelFactory and StreamWriter can load symbol into the main
program instead of the shared libraries specified by the package
information.

To use this feature, the package must be empty in the vpz::Dynamics or
vpz::Output object. Then the library is used to define the symbol to
load. For example:

- vpz::Dynamics { Package = "toto", Libbrary = "titi" } loads the shared
  library $VLE_HOME/pkgs-1.x/toto/lib/simulator/libtiti.so and try to
  found the vle_make_dynamics symbol.

- vpz::Dynamics { Package = "", library="make_titi" } loads the symbol
  "make_titi".

If the library starts with 'exe_' or 'wra_' then symbols respectively
define a vle::devs::Executive or a vle::devs::DynamicsWrapper.