devs: improves observation mechanism
The Observation and Debug mode now is dynamical. In the VPZ, we can
provides a debug="yes" attribute to debug element. The DynamicsDbg is
remove from public API. The DynamicsDbg is now a class between Simulator
and user's dynamics model.

The Observation mechanism use the same behaviour: a class between
Simulator and users'r dynamics or debug model. We adjust the VPZ to
provide observation for output, internal, external, confluent and finish
functions.