vpz: use unique_ptr in vle::vpz::Model to avoid delete
With the previous version of VLE, developers and users must delete the
vpz.project.model.model to avoid memory leaks. Indeed, the
vle::vpz::Model class stores a pointer to the graph but this class is
used in many place in XML parser and in vle.project and vle.classes. To
avoid the problem, we store a unique_ptr and a pointer.

- The unique_ptr is used to manage (in memory) global graph
  (vpz.project.model and vpz.project.classes).
  vle::devs::RootCoordinator use std::move to get the graph.
- The pointer is only used when building the graph structure in XML
  parser.

However, the VPZ parser must be rewrite to reduce and simplify public
API.