utils: download remote package file into temporary file
The remote `package.pkg` file is downloaded into the
`/tmp/package.pkg` file. If a previous file already exists and/or an
error occurred, the previous file is copied and may introduced error.

This patch replaces the `package.pkg` file with a generated file via
the `vle::utils::Path::unique_path` function. (closes: #282).