gvle: fix the build and the install of the project
When using the build or the install action or
the Ctrl+G and/or Ctrl+B, the gui could be locked.
This was due to too much concurrent process launch from
the spawn. Now we garantee one launch at the time,
by disabling the menu and the action also.
By the way the timer usage policy has been changed.
There is one timer per action, an they stay permanent.