cli: add condition parameter to update condition's value
Available with --condition or -c and with an argument following the
syntax:

> vle-2.0  ... -c condition.port=10

Assume condition,port have only a integer value that is force to 10.

> vle-2.0  ... -c condition.port.x=hello

The key 'x' of the map in condition.port must be a string and it will be
assigned to 'hello'.

> vle-2.0  ... -c condition.port.y.1=world

The key 'y' of the map in condition.port must be a set and the index 1
of the vector must be a string and it will be assigned to 'world'.

> vle-2.0  ... -c condition.port.y.2=!

The key 'y' of the map in condition.port must be a set and the index 2
of the vector must be a string and it will be assigned to '!'.

> vle-2.0  ... -c condition.port=5

Assume condition,port have only a real value that is force to 5.0.